Extract methods into files.
